requirements.txt
# ============================================
# AURIX AI ASSISTANT — DEPENDENCIES
# ============================================

# --- Core AI ---
openai>=1.14.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20

# --- Voice ---
openai-whisper>=20231117
TTS>=0.22.0
sounddevice>=0.4.6
soundfile>=0.12.1
numpy>=1.24.0
webrtcvad>=2.0.10
pvporcupine>=3.0.0

# --- System Control ---
pyautogui>=0.9.54
psutil>=5.9.0
screeninfo>=0.8.1

# --- Internet ---
google-search-results>=2.4.2
requests>=2.31.0
beautifulsoup4>=4.12.0
fpdf2>=2.7.0

# --- Memory ---
chromadb>=0.4.22
sentence-transformers>=2.3.0

# --- Vision ---
opencv-python>=4.9.0
face-recognition>=1.3.0
dlib>=19.24.0
Pillow>=10.2.0

# --- Security ---
bcrypt>=4.1.0
pyjwt>=2.8.0

# --- API Server ---
fastapi>=0.109.0
uvicorn>=0.27.0
websockets>=12.0
python-multipart>=0.0.6

# --- Utilities ---
python-dotenv>=1.0.0
pydantic>=2.5.0
rich>=13.7.0
schedule>=1.2.0
aiohttp>=3.9.0
